+++ src/it/scala/au/com/agl/arc/execute/KafkaCommitExecuteSuite.scala
@@ -25,7 +25,7 @@ class KafkaCommitExecuteSuite extends FunSuite with BeforeAndAfter {
   val inputView1 = "inputView1"
   val outputView = "outputView"
   val bootstrapServers = "localhost:29092"
-  val timeout = 3000L
+  val timeout = 1000L
 
   before {
     implicit val spark = SparkSession
@@ -60,6 +60,8 @@ class KafkaCommitExecuteSuite extends FunSuite with BeforeAndAfter {
       .withColumn("normal", randn(seed=27))
       .repartition(10)
       .toJSON
+      .select(col("value").cast(BinaryType))
+
     dataset0.createOrReplaceTempView(inputView0)
     load.KafkaLoad.load(
       KafkaLoad(
@@ -91,14 +93,12 @@ class KafkaCommitExecuteSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty,
-        keyType=StringType,
-        valueType=StringType
+        params=Map.empty
       )
     ).get
 
     var expected = dataset0
-    var actual = extractDataset0.select($"value").as[String]
+    var actual = extractDataset0.select("value")
     assert(actual.except(expected).count === 0)
     assert(expected.except(actual).count === 0)
 
@@ -117,14 +117,12 @@ class KafkaCommitExecuteSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty,
-        keyType=StringType,
-        valueType=StringType
+        params=Map.empty
       )
     ).get
 
     expected = dataset0
-    actual = extractDataset1.select($"value").as[String]
+    actual = extractDataset1.select("value")
     assert(actual.except(expected).count === 0)
     assert(expected.except(actual).count === 0)    
 
@@ -155,12 +153,10 @@ class KafkaCommitExecuteSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty,
-        keyType=StringType,
-        valueType=StringType
+        params=Map.empty
       )
     ).get
-    actual = extractDataset2.select($"value").as[String]
+    actual = extractDataset2.select("value")
     assert(actual.count === 0)
 
     // insert 200 records
@@ -170,6 +166,8 @@ class KafkaCommitExecuteSuite extends FunSuite with BeforeAndAfter {
       .withColumn("normal", randn(seed=27))
       .repartition(10)
       .toJSON
+      .select(col("value").cast(BinaryType))
+
     dataset1.createOrReplaceTempView(inputView1)
     load.KafkaLoad.load(
       KafkaLoad(
@@ -201,14 +199,12 @@ class KafkaCommitExecuteSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty,
-        keyType=StringType,
-        valueType=StringType
+        params=Map.empty
       )
     ).get
 
     expected = dataset1
-    actual = extractDataset3.select($"value").as[String]
+    actual = extractDataset3.select("value")
     assert(actual.except(expected).count === 0)
     assert(expected.except(actual).count === 0)
   }