+++ src/it/scala/au/com/agl/arc/extract/AvroExtractSuite.scala
@@ -1,6 +1,11 @@
 package au.com.agl.arc
 
 import java.net.URI
+import org.apache.commons.lang.StringEscapeUtils
+
+import org.apache.http.client.methods.{HttpPost}
+import org.apache.http.impl.client.HttpClients
+import org.apache.http.entity.StringEntity
 
 import scala.io.Source
 
@@ -16,6 +21,9 @@ import org.apache.spark.sql.functions._
 import au.com.agl.arc.api._
 import au.com.agl.arc.api.API._
 import au.com.agl.arc.util.log.LoggerFactory 
+import au.com.agl.arc.util.ConfigUtils._
+
+import com.typesafe.config._
 
 import au.com.agl.arc.util._
 
@@ -24,7 +32,7 @@ class AvroExtractSuite extends FunSuite with BeforeAndAfter {
   var session: SparkSession = _  
   val targetFile = getClass.getResource("/avro/users.avro").toString 
   val targetBinaryFile = getClass.getResource("/avro/users.avrobinary").toString 
-  val schemaFile = getClass.getResource("/avro/user.avsc").toString 
+  val schemaFile = getClass.getResource("/avro/user.avsc").toString
   val outputView = "dataset"
 
   before {
@@ -76,7 +84,7 @@ class AvroExtractSuite extends FunSuite with BeforeAndAfter {
     implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)  
 
 
-    val schema = new Schema.Parser().parse(CloudUtils.getTextBlob(schemaFile))
+    val schema = new Schema.Parser().parse(CloudUtils.getTextBlob(new URI(schemaFile)))
 
     extract.BytesExtract.extract(
       BytesExtract(
@@ -110,8 +118,122 @@ class AvroExtractSuite extends FunSuite with BeforeAndAfter {
         inputField=None
       )
     ).get
-    
 
-    spark.sql(s"""SELECT value.* FROM ${outputView}""").show(false)
-  }    
+    assert(extractDataset.first.getString(0) == "Alyssa")
+  }   
+
+  test("AvroExtract: Binary with Kafka Schema Registry") {
+    implicit val spark = session
+    implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit val arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false, ignoreEnvironments=false)
+
+
+    val schema = new Schema.Parser().parse(CloudUtils.getTextBlob(new URI(schemaFile)))
+    
+    // POST the schema to kafka so we know it is there as ID=1
+    val httpClient = HttpClients.createDefault
+    val httpPost = new HttpPost("http://localhost:8081/subjects/Kafka-value/versions");
+    httpPost.setEntity(new StringEntity(s"""{"schema": "${StringEscapeUtils.escapeJavaScript(schema.toString)}"}"""));
+    httpPost.addHeader("Content-Type", "application/vnd.schemaregistry.v1+json") 
+    val response = httpClient.execute(httpPost)
+    assert(response.getStatusLine.getStatusCode == 200)
+    response.close
+    httpClient.close
+
+    val conf = s"""{
+      "stages": [
+        {
+          "type": "BytesExtract",
+          "name": "get the binary avro file without header",
+          "environments": [
+            "production",
+            "test"
+          ],
+          "inputURI": "${targetBinaryFile}",
+          "outputView": "bytes_extract_output"
+        },
+        {
+          "type": "AvroExtract",
+          "name": "try to parse",
+          "description": "load customer avro extract",
+          "environments": [
+            "production",
+            "test"
+          ],
+          "inputView": "bytes_extract_output",
+          "outputView": "avro_extract_output",
+          "persist": false,
+          "inputField": "value",
+          "avroSchemaURI": "http://localhost:8081/schemas/ids/1"
+        }        
+      ]
+    }"""
+
+    val base = ConfigFactory.load()
+    val etlConf = ConfigFactory.parseString(conf, ConfigParseOptions.defaults().setSyntax(ConfigSyntax.CONF))
+    val config = etlConf.withFallback(base)
+    var argsMap = collection.mutable.Map[String, String]()
+    val pipeline = ConfigUtils.readPipeline(config.resolve(), "", new URI(""), argsMap, ConfigUtils.Graph(Nil, Nil, false), arcContext)    
+
+    pipeline match {
+      case Left(_) => {
+        println(pipeline)  
+        assert(false)
+      }
+      case Right((pl, _)) => {
+        ARC.run(pl)
+      }
+    }
+  }
+
+  test("AvroExtract: Binary with user.avsc") {
+    implicit val spark = session
+    implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
+    implicit val arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false, ignoreEnvironments=false)
+
+    val conf = s"""{
+      "stages": [
+        {
+          "type": "BytesExtract",
+          "name": "get the binary avro file without header",
+          "environments": [
+            "production",
+            "test"
+          ],
+          "inputURI": "${targetBinaryFile}",
+          "outputView": "bytes_extract_output"
+        },
+        {
+          "type": "AvroExtract",
+          "name": "try to parse",
+          "description": "load customer avro extract",
+          "environments": [
+            "production",
+            "test"
+          ],
+          "inputView": "bytes_extract_output",
+          "outputView": "avro_extract_output",
+          "persist": false,
+          "inputField": "value",
+          "avroSchemaURI": "${schemaFile}"
+        }        
+      ]
+    }"""
+
+    val base = ConfigFactory.load()
+    val etlConf = ConfigFactory.parseString(conf, ConfigParseOptions.defaults().setSyntax(ConfigSyntax.CONF))
+    val config = etlConf.withFallback(base)
+    var argsMap = collection.mutable.Map[String, String]()
+    val pipeline = ConfigUtils.readPipeline(config.resolve(), "", new URI(""), argsMap, ConfigUtils.Graph(Nil, Nil, false), arcContext)    
+
+    pipeline match {
+      case Left(_) => {
+        println(pipeline)  
+        assert(false)
+      }
+      case Right((pl, _)) => {
+        ARC.run(pl)
+      }
+    }
+  }
 }