+++ src/it/scala/au/com/agl/arc/extract/KafkaExtractSuite.scala
@@ -24,7 +24,7 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
   val inputView = "inputView"
   val outputView = "outputView"
   val bootstrapServers = "localhost:29092"
-  val timeout = 3000L
+  val timeout = 1000L
   val checkPointPath = "/tmp/checkpoint"
 
   before {
@@ -47,76 +47,6 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
     session.stop()
     FileUtils.deleteQuietly(new java.io.File(checkPointPath)) 
   }
-
-  test("KafkaExtract: String") {
-    implicit val spark = session
-    import spark.implicits._
-    implicit val logger = LoggerFactory.getLogger(spark.sparkContext.applicationId)
-    implicit val arcContext = ARCContext(jobId=None, jobName=None, environment="test", environmentId=None, configUri=None, isStreaming=false, ignoreEnvironments=false)
-
-    val topic = UUID.randomUUID.toString
-    val groupId = UUID.randomUUID.toString
-
-    // insert 100 records
-    val dataset = spark.sqlContext.range(0, 100)
-      .select("id")
-      .withColumn("uniform", rand(seed=10))
-      .withColumn("normal", randn(seed=27))
-      .repartition(10)
-      .toJSON
-
-    dataset.createOrReplaceTempView(inputView)
-
-    load.KafkaLoad.load(
-      KafkaLoad(
-        name="df", 
-        description=None,
-        inputView=inputView, 
-        topic=topic,
-        bootstrapServers=bootstrapServers,
-        acks= -1,
-        numPartitions=None, 
-        batchSize=16384, 
-        retries=0, 
-        params=Map.empty
-      )
-    )   
-
-    val extractDataset = extract.KafkaExtract.extract(
-      KafkaExtract(
-        name="df", 
-        description=None,
-        outputView=outputView, 
-        topic=topic,
-        bootstrapServers=bootstrapServers,
-        groupID=groupId,
-        maxPollRecords=10000, 
-        timeout=timeout, 
-        autoCommit=false, 
-        persist=true, 
-        numPartitions=None, 
-        partitionBy=Nil,
-        params=Map.empty,
-        keyType=StringType,
-        valueType=StringType
-      )
-    ).get
-
-    val expected = dataset
-    val actual = extractDataset.select(col("value")).as[String]
-
-    val actualExceptExpectedCount = actual.except(expected).count
-    val expectedExceptActualCount = expected.except(actual).count
-    if (actualExceptExpectedCount != 0 || expectedExceptActualCount != 0) {
-      println("actual")
-      actual.show(false)
-      println("expected")
-      expected.show(false)  
-    }
-    assert(actual.except(expected).count === 0)
-    assert(expected.except(actual).count === 0)
-  }  
-
 
   test("KafkaExtract: Binary") {
     implicit val spark = session
@@ -167,9 +97,7 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty,
-        keyType=BinaryType,
-        valueType=BinaryType
+        params=Map.empty
       )
     ).get
 
@@ -204,6 +132,8 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
       .withColumn("normal", randn(seed=27))
       .repartition(10)
       .toJSON
+      .select(col("value").cast(BinaryType))
+
     dataset.createOrReplaceTempView(inputView)
 
     load.KafkaLoad.load(
@@ -235,9 +165,7 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty,
-        keyType=StringType,
-        valueType=StringType        
+        params=Map.empty     
       )
     ).get
 
@@ -255,9 +183,7 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty,
-        keyType=StringType,
-        valueType=StringType   
+        params=Map.empty 
       )
     ).get
 
@@ -294,6 +220,8 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
       .withColumn("normal", randn(seed=27))
       .repartition(10)
       .toJSON
+      .select(col("value").cast(BinaryType))
+
     dataset.createOrReplaceTempView(inputView)
 
     load.KafkaLoad.load(
@@ -325,9 +253,7 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty,
-        keyType=StringType,
-        valueType=StringType   
+        params=Map.empty  
       )
     ).get
 
@@ -345,9 +271,7 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty,
-        keyType=StringType,
-        valueType=StringType   
+        params=Map.empty
       )
     ).get
 
@@ -382,9 +306,7 @@ class KafkaExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true, 
         numPartitions=None, 
         partitionBy=Nil,
-        params=Map.empty,
-        keyType=StringType,
-        valueType=StringType   
+        params=Map.empty
       )
     ).get