+++ src/main/scala/au/com/agl/arc/extract/KafkaExtract.scala
@@ -20,11 +20,16 @@ import au.com.agl.arc.api._
 import au.com.agl.arc.api.API._
 import au.com.agl.arc.util._
 
-object KafkaExtract {
+case class KafkaRecord (
+  topic: String,
+  partition: Int,
+  offset: Long,
+  timestamp: Long,
+  key: Array[Byte],
+  value: Array[Byte]
+)
 
-  /** Phantom Type to enable compiler to find the encoder we want
-    */
-  type KafkaRow = Row  
+object KafkaExtract {
 
   def extract(extract: KafkaExtract)(implicit spark: SparkSession, logger: au.com.agl.arc.util.log.logger.Logger, arcContext: ARCContext): Option[DataFrame] = {
     import spark.implicits._
@@ -57,38 +62,22 @@ object KafkaExtract {
         .option("subscribe", extract.topic)
         .load()
     } else {
-      // dynamically create schema
-      val baseSchema = List(
-        StructField("topic", StringType, false),
-        StructField("partition", IntegerType, false),
-        StructField("offset", LongType, false),
-        StructField("timestamp", LongType, false)
-      )
-
-      val extractSchema = (extract.keyType, extract.valueType) match {
-        case (StringType, StringType) => StructType(baseSchema ::: List(StructField("key", StringType, true), StructField("value", StringType, false)))
-        case (BinaryType, StringType) => StructType(baseSchema ::: List(StructField("key", BinaryType, true), StructField("value", StringType, false)))
-        case (StringType, BinaryType) => StructType(baseSchema ::: List(StructField("key", StringType, true), StructField("value", BinaryType, false)))
-        case (BinaryType, BinaryType) => StructType(baseSchema ::: List(StructField("key", BinaryType, true), StructField("value", BinaryType, false)))
-      }
-
-      /** Create a dynamic RowEncoder from the provided schema. We use the phantom
-        * TypeRow type to enable implicit resolution to find our encoder.
-        */
-      implicit val typedEncoder: Encoder[KafkaRow] = org.apache.spark.sql.catalyst.encoders.RowEncoder(extractSchema)
-          
       // KafkaConsumer properties
       // https://kafka.apache.org/documentation/#consumerconfigs
-      val props = new Properties
-      props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, extract.bootstrapServers)
-      props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer")
-      props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer")
-      props.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, "false")
-      props.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest")
-      props.put(ConsumerConfig.REQUEST_TIMEOUT_MS_CONFIG, extract.timeout.toString)
-      props.put(ConsumerConfig.SESSION_TIMEOUT_MS_CONFIG, Math.min(10000, extract.timeout-1).toString)
-      props.put(ConsumerConfig.FETCH_MAX_WAIT_MS_CONFIG, Math.min(500, extract.timeout-1).toString)
-      props.put(ConsumerConfig.HEARTBEAT_INTERVAL_MS_CONFIG, Math.min(3000, extract.timeout-2).toString)
+
+
+      val baseProps = new Properties
+      baseProps.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, extract.bootstrapServers)
+      baseProps.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.ByteArrayDeserializer")
+      baseProps.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.ByteArrayDeserializer")
+      baseProps.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, "false")
+      baseProps.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest")
+      baseProps.put(ConsumerConfig.REQUEST_TIMEOUT_MS_CONFIG, extract.timeout.toString)
+      baseProps.put(ConsumerConfig.SESSION_TIMEOUT_MS_CONFIG, Math.min(10000, extract.timeout-1).toString)
+      baseProps.put(ConsumerConfig.FETCH_MAX_WAIT_MS_CONFIG, Math.min(500, extract.timeout-1).toString)
+      baseProps.put(ConsumerConfig.HEARTBEAT_INTERVAL_MS_CONFIG, Math.min(3000, extract.timeout-2).toString)
+
+      val props = baseProps
       props.put(ConsumerConfig.GROUP_ID_CONFIG, extract.groupID)
 
       // first get the number of partitions via the driver process so it can be used for mapPartition
@@ -106,58 +95,26 @@ object KafkaExtract {
       }
 
       try {
-        spark.sqlContext.emptyDataFrame.repartition(numPartitions).mapPartitions[KafkaRow] { partition: Iterator[Row] => 
+        spark.sqlContext.emptyDataFrame.repartition(numPartitions).mapPartitions(partition => {
           // get the partition of this executor which maps 1:1 with Kafka partition
           val partitionId = TaskContext.getPartitionId
-
-          val props = new Properties
-          props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, extract.bootstrapServers)
-
-          (extract.keyType, extract.valueType) match {
-            case (StringType, StringType) => {
-              props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer")
-              props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer")
-            }
-            case (BinaryType, StringType) => {
-              props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.ByteArrayDeserializer")
-              props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer")
-            }
-            case (StringType, BinaryType) => {
-              props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer")
-              props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.ByteArrayDeserializer")
-            }
-            case (BinaryType, BinaryType) => {
-              props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.ByteArrayDeserializer")
-              props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.ByteArrayDeserializer")
-            }
-          }
-
-          props.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, "false")
-          props.put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest")
-          props.put(ConsumerConfig.REQUEST_TIMEOUT_MS_CONFIG, extract.timeout.toString)
-          props.put(ConsumerConfig.SESSION_TIMEOUT_MS_CONFIG, Math.min(10000, extract.timeout-1).toString)
-          props.put(ConsumerConfig.FETCH_MAX_WAIT_MS_CONFIG, Math.min(500, extract.timeout-1).toString)
-          props.put(ConsumerConfig.HEARTBEAT_INTERVAL_MS_CONFIG, Math.min(3000, extract.timeout-2).toString)
+          
+          val props = baseProps
           props.put(ConsumerConfig.MAX_POLL_RECORDS_CONFIG, extract.maxPollRecords.toString)
           props.put(ConsumerConfig.GROUP_ID_CONFIG, s"${extract.groupID}-${partitionId}")
 
           // try to assign records based on partitionId and extract 
-          val kafkaConsumer = (extract.keyType, extract.valueType) match {
-            case (StringType, StringType) => new KafkaConsumer[String, String](props)
-            case (BinaryType, StringType) => new KafkaConsumer[Array[Byte], String](props)
-            case (StringType, BinaryType) => new KafkaConsumer[String, Array[Byte]](props)
-            case (BinaryType, BinaryType) => new KafkaConsumer[Array[Byte], Array[Byte]](props)
-          }
+          val kafkaConsumer = new KafkaConsumer[Array[Byte], Array[Byte]](props)
           val topicPartition = new TopicPartition(extract.topic, partitionId)
 
-          def getKafkaRecord(): List[KafkaRow] = {
+          def getKafkaRecord(): List[KafkaRecord] = {
             kafkaConsumer.poll(extract.timeout).records(extract.topic).asScala.map(consumerRecord => {
-              Row.fromSeq(Seq(consumerRecord.topic, consumerRecord.partition, consumerRecord.offset, consumerRecord.timestamp, consumerRecord.key, consumerRecord.value)).asInstanceOf[KafkaRow]
+              KafkaRecord(consumerRecord.topic, consumerRecord.partition, consumerRecord.offset, consumerRecord.timestamp, consumerRecord.key, consumerRecord.value)
             }).toList
           }
 
           @tailrec
-          def getAllKafkaRecords(kafkaRecords: List[KafkaRow], kafkaRecordsAccumulator: List[KafkaRow]): List[KafkaRow] = {
+          def getAllKafkaRecords(kafkaRecords: List[KafkaRecord], kafkaRecordsAccumulator: List[KafkaRecord]): List[KafkaRecord] = {
               kafkaRecords match {
                   case Nil => kafkaRecordsAccumulator
                   case _ => getAllKafkaRecords(getKafkaRecord, kafkaRecordsAccumulator ::: kafkaRecords)
@@ -179,7 +136,7 @@ object KafkaExtract {
           } finally {
             kafkaConsumer.close
           }
-        }.toDF
+        }).toDF
       } catch {
         case e: Exception => throw new Exception(e) with DetailException {
           override val detail = stageDetail