+++ src/main/scala/au/com/agl/arc/util/ConfigUtils.scala
@@ -50,12 +50,12 @@ object ConfigUtils {
   def getConfigString(uri: URI, argsMap: collection.mutable.Map[String, String], arcContext: ARCContext)(implicit spark: SparkSession, logger: au.com.agl.arc.util.log.logger.Logger): Either[List[Error], String] = {
     uri.getScheme match {
       case "local" => {
-        val filePath = SparkFiles.get(uri.getPath)
+        val filePath = new URI(SparkFiles.get(uri.getPath))
         val etlConfString = CloudUtils.getTextBlob(filePath)
         Right(etlConfString)
       }
       case "file" => {
-        val etlConfString = CloudUtils.getTextBlob(uri.toString)
+        val etlConfString = CloudUtils.getTextBlob(uri)
         Right(etlConfString)
       }
       // amazon s3
@@ -85,7 +85,7 @@ object ConfigUtils {
         val config = Map("fs_s3a_endpoint" -> endpoint, "fs_s3a_connection_ssl_enabled" -> connectionSSLEnabled, "fs_s3a_access_key" -> accessKey, "fs_s3a_secret_key" -> secretKey)
 
         CloudUtils.setHadoopConfiguration(Some(Authentication.AmazonAccessKey(accessKey, secretKey)))
-        val etlConfString = CloudUtils.getTextBlob(uri.toString)
+        val etlConfString = CloudUtils.getTextBlob(uri)
         Right(etlConfString)
       }
       // azure blob
@@ -103,7 +103,7 @@ object ConfigUtils {
         }
 
         CloudUtils.setHadoopConfiguration(Some(Authentication.AzureSharedKey(accountName, accountKey)))
-        val etlConfString = CloudUtils.getTextBlob(uri.toString)
+        val etlConfString = CloudUtils.getTextBlob(uri)
         Right(etlConfString)
       }
       // azure data lake storage
@@ -121,7 +121,7 @@ object ConfigUtils {
         }
 
         CloudUtils.setHadoopConfiguration(Some(Authentication.AzureDataLakeStorageToken(clientID, refreshToken)))
-        val etlConfString = CloudUtils.getTextBlob(uri.toString)
+        val etlConfString = CloudUtils.getTextBlob(uri)
         Right(etlConfString)
       }
       // google cloud
@@ -139,7 +139,7 @@ object ConfigUtils {
         }
 
         CloudUtils.setHadoopConfiguration(Some(Authentication.GoogleCloudStorageKeyFile(projectID, keyFilePath)))
-        val etlConfString = CloudUtils.getTextBlob(uri.toString)
+        val etlConfString = CloudUtils.getTextBlob(uri)
         Right(etlConfString)
       }
       case "classpath" => {
@@ -750,6 +750,24 @@ object ConfigUtils {
     }
   }
 
+  private def parseAvroSchema(path: String, schemaString: String)(implicit c: Config): Either[Errors, Option[org.apache.avro.Schema]] = {
+    def err(lineNumber: Option[Int], msg: String): Either[Errors, Option[org.apache.avro.Schema]] = Left(ConfigError(path, lineNumber, msg) :: Nil)
+    try {
+      // if schema contains backslash it might have come from the kafka schema registry therefore try to get the data out of the returned object
+      // this is hacky but needs to behave well with a registry response
+      if (schemaString.contains(""""schema":""") && schemaString.contains("\\")) {
+        val objectMapper = new ObjectMapper()
+        val metaTree = objectMapper.readTree(schemaString)
+        Right(Option(new org.apache.avro.Schema.Parser().parse(metaTree.get("schema").asText)))
+      } else {
+        // try to parse schema
+        Right(Option(new org.apache.avro.Schema.Parser().parse(schemaString)))
+      }
+    } catch {
+      case e: Exception => err(Some(c.getValue(path).origin.lineNumber()), e.getMessage)
+    }
+  }
+
   private def parseGlob(path: String, glob: String)(implicit c: Config): Either[Errors, String] = {
     def err(lineNumber: Option[Int], msg: String): Either[Errors, String] = Left(ConfigError(path, lineNumber, msg) :: Nil)
 
@@ -781,7 +799,7 @@ object ConfigUtils {
     def err(lineNumber: Option[Int], msg: String): Either[Errors, String] = Left(ConfigError(path, lineNumber, msg) :: Nil)
 
     try {
-      val textFile = CloudUtils.getTextBlob(uri.toString)
+      val textFile = CloudUtils.getTextBlob(uri)
       if (textFile.length == 0) {
         err(Some(c.getValue(path).origin.lineNumber()), s"""File at ${uri.toString} is empty.""")
       } else {
@@ -857,7 +875,7 @@ object ConfigUtils {
   def readAvroExtract(idx: Int, graph: Graph, name: StringConfigValue, params: Map[String, String])(implicit spark: SparkSession, logger: au.com.agl.arc.util.log.logger.Logger, c: Config): (Either[List[StageError], PipelineStage], Graph) = {
     import ConfigReader._
 
-    val expectedKeys = "type" :: "name" :: "description" :: "environments" :: "inputURI" :: "inputView" :: "outputView" :: "authentication" :: "contiguousIndex" :: "numPartitions" :: "partitionBy" :: "persist" :: "schemaURI" :: "schemaView" :: "params" :: "basePath" :: "inputField" :: Nil
+    val expectedKeys = "type" :: "name" :: "description" :: "environments" :: "inputURI" :: "inputView" :: "outputView" :: "authentication" :: "contiguousIndex" :: "numPartitions" :: "partitionBy" :: "persist" :: "schemaURI" :: "schemaView" :: "params" :: "basePath" :: "avroSchemaURI" :: "inputField" :: Nil
     val invalidKeys = checkValidKeys(c)(expectedKeys)
 
     val description = getOptionalValue[String]("description")
@@ -889,16 +907,28 @@ object ConfigUtils {
     val basePath = getOptionalValue[String]("basePath")
     val inputField = getOptionalValue[String]("inputField")
 
-    (name, description, extractColumns, schemaView, input, parsedGlob, outputView, persist, numPartitions, partitionBy, authentication, contiguousIndex, invalidKeys, basePath, inputField) match {
-      case (Right(n), Right(d), Right(cols), Right(sv), Right(in), Right(pg), Right(ov), Right(p), Right(np), Right(pb), Right(auth), Right(ci), Right(_), Right(bp), Right(ipf)) =>
+    val avroSchemaURI = getOptionalValue[String]("avroSchemaURI")
+    val avroSchema: Either[Errors, Option[org.apache.avro.Schema]] = avroSchemaURI.rightFlatMap(optAvroSchemaURI => 
+      optAvroSchemaURI match { 
+        case Some(uri) => {
+          parseURI("avroSchemaURI", uri)
+          .rightFlatMap(uri => textContentForURI(uri, "avroSchemaURI", Right(None) ))
+          .rightFlatMap(schemaString => parseAvroSchema("avroSchemaURI", schemaString))
+        }
+        case None => Right(None)
+      }
+    )    
+
+    (name, description, extractColumns, schemaView, input, parsedGlob, outputView, persist, numPartitions, partitionBy, authentication, contiguousIndex, invalidKeys, basePath, inputField, avroSchemaURI, avroSchema) match {
+      case (Right(n), Right(d), Right(cols), Right(sv), Right(in), Right(pg), Right(ov), Right(p), Right(np), Right(pb), Right(auth), Right(ci), Right(_), Right(bp), Right(ipf), Right(_), Right(avsc)) =>
         val input = if(c.hasPath("inputView")) Left(in) else Right(pg)
         val schema = if(c.hasPath("schemaView")) Left(sv) else Right(cols)
 
         var outputGraph = graph.addVertex(Vertex(idx, ov))
 
-        (Right(AvroExtract(n, d, schema, ov, input, auth, params, p, np, pb, ci, bp, None, ipf)), outputGraph)
+        (Right(AvroExtract(n, d, schema, ov, input, auth, params, p, np, pb, ci, bp, avsc, ipf)), outputGraph)
       case _ =>
-        val allErrors: Errors = List(name, description, input, schemaView, parsedGlob, outputView, persist, numPartitions, partitionBy, authentication, contiguousIndex, extractColumns, invalidKeys, basePath, inputField).collect{ case Left(errs) => errs }.flatten
+        val allErrors: Errors = List(name, description, input, schemaView, parsedGlob, outputView, persist, numPartitions, partitionBy, authentication, contiguousIndex, extractColumns, invalidKeys, basePath, inputField, avroSchemaURI, avroSchema).collect{ case Left(errs) => errs }.flatten
         val stageName = stringOrDefault(name, "unnamed stage")
         val err = StageError(idx, stageName, c.origin.lineNumber, allErrors)
         (Left(err :: Nil), graph)
@@ -1261,7 +1291,7 @@ object ConfigUtils {
   def readKafkaExtract(idx: Int, graph: Graph, name: StringConfigValue, params: Map[String, String])(implicit spark: SparkSession, logger: au.com.agl.arc.util.log.logger.Logger, c: Config): (Either[List[StageError], PipelineStage], Graph) = {
     import ConfigReader._
 
-    val expectedKeys = "type" :: "name" :: "description" :: "environments" :: "outputView" :: "bootstrapServers" :: "topic" :: "groupID" :: "autoCommit" :: "maxPollRecords" :: "numPartitions" :: "partitionBy" :: "persist" :: "timeout" :: "params" :: "keyType" :: "valueType" :: Nil
+    val expectedKeys = "type" :: "name" :: "description" :: "environments" :: "outputView" :: "bootstrapServers" :: "topic" :: "groupID" :: "autoCommit" :: "maxPollRecords" :: "numPartitions" :: "partitionBy" :: "persist" :: "timeout" :: "params" :: Nil
     val invalidKeys = checkValidKeys(c)(expectedKeys)    
 
     val description = getOptionalValue[String]("description")
@@ -1279,16 +1309,12 @@ object ConfigUtils {
     val timeout = getValue[Long]("timeout", default = Some(10000L))
     val autoCommit = getValue[Boolean]("autoCommit", default = Some(false))
 
-
-    val keyType = getValue[String]("keyType", default = Some("string"), validValues = "string" :: "binary" :: Nil) |> parseDataType("keyType") _
-    val valueType = getValue[String]("valueType", default = Some("string"), validValues = "string" :: "binary" :: Nil) |> parseDataType("valueType") _
-
-    (name, description, outputView, topic, bootstrapServers, groupID, persist, numPartitions, maxPollRecords, timeout, autoCommit, partitionBy, invalidKeys, keyType, valueType) match {
-      case (Right(n), Right(d), Right(ov), Right(t), Right(bs), Right(g), Right(p), Right(np), Right(mpr), Right(time), Right(ac), Right(pb), Right(_), Right(kt), Right(vt)) => 
+    (name, description, outputView, topic, bootstrapServers, groupID, persist, numPartitions, maxPollRecords, timeout, autoCommit, partitionBy, invalidKeys) match {
+      case (Right(n), Right(d), Right(ov), Right(t), Right(bs), Right(g), Right(p), Right(np), Right(mpr), Right(time), Right(ac), Right(pb), Right(_)) => 
         var outputGraph = graph.addVertex(Vertex(idx, ov))
-        (Right(KafkaExtract(n, d, ov, t, bs, g, mpr, time, ac, params, p, np, pb, kt, vt)), outputGraph)
+        (Right(KafkaExtract(n, d, ov, t, bs, g, mpr, time, ac, params, p, np, pb)), outputGraph)
       case _ =>
-        val allErrors: Errors = List(name, description, outputView, topic, bootstrapServers, groupID, persist, numPartitions, maxPollRecords, timeout, autoCommit, partitionBy, invalidKeys, keyType, valueType).collect{ case Left(errs) => errs }.flatten
+        val allErrors: Errors = List(name, description, outputView, topic, bootstrapServers, groupID, persist, numPartitions, maxPollRecords, timeout, autoCommit, partitionBy, invalidKeys).collect{ case Left(errs) => errs }.flatten
         val stageName = stringOrDefault(name, "unnamed stage")
         val err = StageError(idx, stageName, c.origin.lineNumber, allErrors)
         (Left(err :: Nil), graph)